soca-dashboard/requirements.txt
django>=5.0
requests>=2.31
python-dotenv>=1.0
ruamel.yaml>=0.18
